refactor(vlm): move the mRoPE get_rope_index resolver to _transformers/utils
The resolver is a transformers-layout concern, not recipe logic, so it
belongs next to the other HuggingFace bridge helpers where other recipes
can reach it. Renamed to resolve_get_rope_index now that it is imported
across modules.
