+++ nemo_automodel/_transformers/utils.py
@@ -13,12 +13,41 @@
 # limitations under the License.
 
 import logging
+from collections.abc import Callable
 from typing import Any, Optional
 
 import torch
+import torch.nn as nn
 from transformers import AutoConfig
 
 logger = logging.getLogger(__name__)
+
+
+def resolve_get_rope_index(model: nn.Module) -> Optional[Callable]:
+    """Locate a model's mRoPE position-id builder.
+
+    Transformers defines ``get_rope_index`` on the base model rather than on the
+    ``*ForConditionalGeneration`` that ``model_parts`` holds (Qwen2-VL,
+    Qwen2.5-VL, Qwen3-VL and Qwen3-VL-MoE all follow this layout), and DDP or
+    MegatronFSDP add another wrapper on top without proxying attribute reads. A
+    plain ``getattr`` on the top-level module therefore finds nothing, and packed
+    multimodal training silently falls back to 1D positions. Unwrap ``module``
+    first, then check the model itself, then HuggingFace's ``base_model``
+    property, which resolves to ``getattr(model, model.base_model_prefix, model)``.
+
+    Args:
+        model: The (possibly wrapped) model to search. Accepts any module; no
+            tensor inputs.
+
+    Returns:
+        The ``get_rope_index`` callable, or ``None`` when the model does not
+        expose one (i.e. it is not an mRoPE model).
+    """
+    model = getattr(model, "module", model)
+    get_rope_index = getattr(model, "get_rope_index", None)
+    if get_rope_index is not None:
+        return get_rope_index
+    return getattr(getattr(model, "base_model", None), "get_rope_index", None)
 
 
 def _should_load_before_shard(
